backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
djangorestframework==3.15.2
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.5.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
